Additional Classes Stacked
Added the class being used for individual links of other sidebar boxes in order to save space, as well as abide by DRY standards.  The original intent was soley to use these additional classes for the layout of individual items, however the hover effects ended up being an added bonus.